fix referenceerror from import
