+++ cli/subwebapp.ts
@@ -1,0 +1,28 @@
+// All of our various create react apps
+export const SUB_WEBAPPS = [
+    {
+        name: "kiosk",
+        buildCss: false,
+        localServe: false
+    },
+    {
+        name: "teachertool",
+        buildCss: false,
+        localServe: false
+    },
+    {
+        name: "skillmap",
+        buildCss: true,
+        localServe: true
+    },
+    {
+        name: "multiplayer",
+        buildCss: true,
+        localServe: true
+    },
+    {
+        name: "authcode",
+        buildCss: true,
+        localServe: true
+    },
+];
